Move utils function back into enrichr.py
